--- src/redux/initialState.ts
@@ -1,21 +1,0 @@
-export const initialState = {
-	character: {
-		startedHp: 100,
-		startedArmor: 0,
-		startedDps: 10,
-		lvl: 1,
-		exp: 0
-	},
-	game: {
-		time: {min: '00', sec: '00'},
-		lvl: 1
-	},
-	user: {
-		mail: null,
-		login: null,
-		name: null,
-		secondName: null,
-		nameInChat: null,
-		phone: null
-	}
-};
