--- src/redux/types.ts
@@ -1,5 +1,0 @@
-import {RouterState} from 'connected-react-router';
-
-export interface State {
-  readonly router: RouterState;
-}
