+++ src/server/server-render-middleware.tsx
@@ -8,7 +8,7 @@ import {renderToStaticMarkup, renderToString} from 'react-dom/server';
 import parse from 'html-react-parser';
 import {App} from '../components/App';
 import {StaticRouter} from 'react-router-dom';
-import {createReduxStore, getInitialState} from '../redux/rootStore';
+import {store} from '../redux/rootStore';
 
 export default (req: Request, res: Response) => {
 	const statsFile = path.resolve('./dist/loadable-stats.json');
@@ -16,7 +16,7 @@ export default (req: Request, res: Response) => {
 
 	const location = req.url;
 	const context: StaticRouterContext = {};
-	const {store} = createReduxStore(getInitialState(location), location);
+
 
 	const appContent = chunkExtractor.collectChunks(
 		<Provider store={store}>
