+++ src/utils/isServerEnvChecker.ts
@@ -1,0 +1,5 @@
+export const isServer = !(
+	typeof window !== 'undefined' &&
+  window.document &&
+  window.document.createElement
+);
